[REF] tests: split checkers
task-2801043

X-original-commit: 8e8c4668927681f364adb8b5904e64bece78fce8
