+++ tests/checkers/__init__.py
@@ -1,0 +1,3 @@
+from . import redirect_rules
+from . import resource_files
+from . import rst_style
